// movinin/api/src/config/env.config.ts
export const MINIMUM_AGE = Number.parseInt(__env__('MI_MINIMUM_AGE', false, '21'), 10)
// ...
export interface Property extends Document {
  name: string
  type: movininTypes.PropertyType
  agency?: Types.ObjectId
  owner?: Types.ObjectId
  description: string
  image: string
  images?: string[]
  bedrooms: number
  bathrooms: number
  kitchens?: number
  parkingSpaces?: number,
  size?: number
  petsAllowed: boolean
  furnished: boolean
  minimumAge: number
  location: Types.ObjectId
  address?: string
  latitude?: number
  longitude?: number
  price: number
  hidden?: boolean
  cancellation?: number
  aircon?: boolean
  available?: boolean
  rentalTerm: movininTypes.RentalTerm
  vrboEnabled?: boolean
  vrboListingId?: string
}